+++ examples/csr-complete/src/lib.rs
@@ -98,7 +98,7 @@ fn render_language(lang: &'static Language) -> impl IntoView {
                 id=lang
                 name="language"
                 value=lang
-                checked=lang.is_active()
+                checked=i18n.language.get() == lang
                 on:click=move |_| i18n.language.set(lang)
                 type="radio"
             />